remove everything
There was a bit of a mess where Trebuchet was living in the plugins directory
of a Rails app while under heavy development. When I pulled the changes back
into the Trebuchet project, I did so very forcefully by deleting the old master
branch and renaming it. The elegant part was using subtree and directory filtering
to extract it, but the graft was not gentle.

This time, I'm choosing to delete all files from this branch, then merge in the
downstream changes since they share no common ancestor. This may preserve a bit
more history, but more importantly is gentler than essentially a force push to
master of a project linked by SHA in many gemfiles.

I hope that history will forgive me.