Raise an error when someone attempts to aim() at a percent based strategy
This change disables new uses of the percent and percent_visited strategies and also introduces the new aliases percent_drecated and percent_visited_deprecated.

The goal is to switch all current uses of the percent based strategies over to the *deprecated alias names after this change goes out. This will enable us to follow up with another commit that corrects the implementation of the percent based strategies without affecting the set of users that are exposed to the deprecated implementation of the strategies.

We hope to re-implement the percent based strategies in a way that can be used safely with experiment analysis. The current implemention tends to distribute features using the percent based strategies over the same subsets of users making it difficult to draw conclusions from the results of the analysis of those features.

Testing:
I wrote new specs and verified that all specs passed