Cache Trebuchet#launch?
Implements a simple cache on Trebuchet#launch?.  This avoids expensive
construction and evaluation of strategies for calls on the same feature.
This also causes the result of Trebuchet#launch? to be fixed for the
lifespan of the Trebuchet instance, which seems desirable.

The Trebuchet instance is created around a user and a request, with
neither of these fields being mutable.  Strategies, including custom
strategies, are functions of (user, request), so this should be a safe
level to cache at.  In test, Trebuchet.visitor_id is global state that
influences the result of some strategies, but in actual usage,
visitor_id is defined as a function of request.

Additionally, this has the side effect of covering over caching behavior
in some backends that results in features for which no strategy has been
defined are not cached (forcing the worst case performance).